Fix the LowPowerMode
When the sketch is idle in loop(), the module's low-power mode will be enabled, which makes the following operations (e.g. after a reboot or re-flashing the sketch) fail, since the module will still be in low-power mode.

Fix this by forcing the module to wake up at the beginning of setup(), and wait 100ms until this seems to have an effect. (This worked 10 out of 10 times when testing with an Arduino Uno on 5V.)