Fix to support all SAMD21 proccessors
There is a issue where some ATSAMD21 boards such as the Adafruit ItsyBitsy that don't work with this library due to line 171 defining the section for an Arduino Zero instead of the processor used on the board. Tested with Adafruit Feather M0 and Adafruit ItsyBitsy M0, Feather worked before the change.